AELIP-15: Fixes and improvements
