+++ public/data/nft-metadata/goerli-metadata.json
@@ -4,11 +4,11 @@
     "address": "0xf5de760f2e916647fd766B4AD9E85ff943cE3A2b",
     "name": "Test ERC-721",
     "slug": "test-erc-721",
-    "imageUrl": "https://storage.googleapis.com/goerli-pixxiti/c7f8510719344f67d90fb82a3bae2161b848171d",
+    "imageUrl": "https://ipfs.io/ipfs/bafybeifvwitulq6elvka2hoqhwixfhgb42l4aiukmtrw335osetikviuuu",
     "isVerified": true,
     "numOwners": 8,
     "totalSupply": 9999,
-    "contractType": "erc-721",
+    "contractType": "erc721",
     "floorPrice": 0.23,
     "totalVolume": 23.1921932853,
     "paymentSymbol": "ETH",
@@ -23,7 +23,7 @@
     "isVerified": true,
     "numOwners": 8,
     "totalSupply": 9999,
-    "contractType": "erc-1155",
+    "contractType": "erc1155",
     "floorPrice": 0.23,
     "totalVolume": 23.1921932853,
     "paymentSymbol": "ETH",
