+++ public/data/nft-metadata/quixotic-metadata.json
